refactor mocks and model WIP
