+++ src/mock/destinations.js
@@ -1,0 +1,59 @@
+export const mockDestinations = [
+  {
+    id: 'pa',
+    description: 'The capital of France, known for its art, culture, and monuments.',
+    name: 'Paris',
+    photos: [
+      {
+        src: 'https://loremflickr.com/248/152?random=1',
+        description: 'The Louvre Museum, located in the city center, is one of the most visited museums in the world.'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=2',
+        description: 'The Louvre Museum, located in the city center, is one of the most visited museums in the world.'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=3',
+        description: 'The Louvre Museum, located in the city center, is one of the most visited museums in the world.'
+      }
+    ],
+  },
+  {
+    id: 'lo',
+    description: 'The capital of Great Britain, known for its art, culture, and monuments.',
+    name: 'London',
+    photos: [
+      {
+        src: 'https://loremflickr.com/248/152?random=4',
+        description: 'London Eye, located in the city center.'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=5',
+        description: 'London Eye, located in the city center.'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=6',
+        description: 'London Eye, located in the city center.'
+      }
+    ],
+  },
+  {
+    id: 'am',
+    description: 'The capital of The Netherlands, known for its art, culture, and monuments.',
+    name: 'Amsterdam',
+    photos: [
+      {
+        src: 'https://loremflickr.com/248/152?random=7',
+        description: 'Red Lights District'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=8',
+        description: 'Red Lights District'
+      },
+      {
+        src: 'https://loremflickr.com/248/152?random=9',
+        description: 'Red Lights District'
+      }
+    ],
+  },
+];
