+++ src/mock/offers.js
@@ -1,0 +1,67 @@
+export const mockOffers = [
+  {
+    type: 'taxi',
+    offers: [
+      {
+        id: 't1',
+        title: 'Upgrade to business class',
+        price: 100,
+      },
+      {
+        id: 't2',
+        title: 'Order Uber',
+        price: 20,
+      }
+    ]
+  },
+  {
+    type: 'flight',
+    offers: [
+      {
+        id: 'f1',
+        title: 'Add laggage',
+        price: 50,
+      },
+      {
+        id: 'f2',
+        title: 'Switch to comfort',
+        price: 80,
+      }
+    ]
+  },
+  {
+    type: 'drive',
+    offers: [
+      {
+        id: 'd1',
+        title: 'Rent a car',
+        price: 200,
+      }
+    ]
+  },
+  {
+    type: 'check-in',
+    offers: [
+      {
+        id: 'ci1',
+        title: 'Add breakfast',
+        price: 50,
+      }
+    ]
+  },
+  {
+    type: 'sightseeing',
+    offers: [
+      {
+        id: 'ss1',
+        title: 'Book tickets',
+        price: 40,
+      },
+      {
+        id: 'ss2',
+        title: 'Lunch in city',
+        price: 30,
+      }
+    ]
+  }
+];
