+++ src/mock/points.js
@@ -1,0 +1,54 @@
+import { getRandomArrayElement } from '../utils/common.js';
+
+export const mockPoints = [
+  {
+    id: 'point1',
+    basePrice: 300,
+    dateFrom: '2024-11-22T08:00:00',
+    dateTo: '2024-11-22T11:30:00',
+    destination: 'pa',
+    isFavorite: true,
+    offers: [
+      'f1',
+      'f2'
+    ],
+    type: 'flight',
+  },
+  {
+    id: 'point2',
+    basePrice: 260,
+    dateFrom: '2024-12-23T14:00:00',
+    dateTo: '2024-12-23T14:30:00',
+    destination: 'lo',
+    isFavorite: false,
+    offers: [
+      't1',
+      't2'
+    ],
+    type: 'taxi',
+  },
+  {
+    id: 'point3',
+    basePrice: 200,
+    dateFrom: '2024-12-26T08:00:00',
+    dateTo: '2024-12-27T22:00:00',
+    destination: 'am',
+    isFavorite: true,
+    offers: [],
+    type: 'check-in',
+  }
+];
+
+// function getOffersByType(type) {
+//   return mockOffers.find((offer) => offer.type === type).offers;
+// }
+
+// function getRandomDestination() {
+//   return getRandomArrayElement(mockDestinations);
+// }
+
+function getRandomPoint() {
+  return getRandomArrayElement(mockPoints);
+}
+
+export { getRandomPoint };
